requirements.txt

numpy==1.22.4
paho-mqtt==1.6.1
pandas==1.4.2
schedule==1.1.0
